Create Update.sh
I noticed you do not have an auto update script for us linux users so I went and created one for you.

It a simple bash file that can take 3 arguments

Update.sh = Checks to see if there is an update, if yes update
Update.sh -update = Same as above
Update.sh -force = Will update regardless if its needed or not
Update.sh -check = Will check for an update, and prompt you if you want to update or not.

---

Because there was not an easy way (that I knew of) to get the current version (maybe have an API or file with the version number) It will one first attempt assume its outdated.

After the initial update, it will create a folder + file in the rdtcs root folder called .installed/version.txt and store the installed version in the text.  It will then start comparing the latest github version to whats in the file.

When it updates, it will check for the latest version using github's API and grab the version and download link from there.  Afterwards, it will then 
1. download the zip file
2. Stop the rdtc service
3. Backup the config and DB to .backup
4. Unzip the latest version
5. Update the version.txt file
6. Restore backup
7. Restart rdtc

Would most likely be best to set a cron to run it on a set schedule either daily or weekly.  If there is no update it will just exit, otherwise its fully automated.